player partialize onupdate
