+++ Assets/Code/Fsm/GravityFsm/PlayerFsm/PlayerFsm_Aerial.cs
@@ -1,0 +1,7 @@
+public partial class PlayerFsm
+{
+    private void AerialOnUpdate()
+    {
+        Animator.SetLayerWeight(1, 0);
+    }
+}
