+++ Assets/Code/Fsm/GravityFsm/PlayerFsm/PlayerFsm_Dashsquat.cs
@@ -1,0 +1,8 @@
+public partial class PlayerFsm
+{
+    private void DashsquatOnUpdate()
+    {
+        HandleCollisionMove();
+        HandleTurning(DashsquatTurnMultiplier, true);
+    }
+}
