+++ Assets/Code/Fsm/GravityFsm/PlayerFsm/PlayerFsm_ForceWallRotation.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 631e86c0cf3f4fc59d04ee3b64174b64
+timeCreated: 1759376057
