+++ Assets/Code/Fsm/GravityFsm/PlayerFsm/PlayerFsm_GrappleStartup.cs
@@ -1,0 +1,24 @@
+using UnityEngine;
+
+public partial class PlayerFsm
+{
+    private void GrappleStartupOnUpdate()
+    {
+        Animator.SetLayerWeight(2, 0);
+        var transformPosition = new Vector3(PlayerWeaponFsm.Singleton.transform.position.x, transform.position.y,
+            PlayerWeaponFsm.Singleton.transform.position.z);
+        var forward = transformPosition - transform.position;
+        transform.rotation = Quaternion.Slerp(transform.rotation, Quaternion.LookRotation(forward, Vector3.up),
+            Time.deltaTime * GrappleStartupRotationLerpStrength);
+
+        var destinationPosition = new Vector3(transform.position.x,
+            PlayerWeaponFsm.Singleton.transform.position.y + GrappleStartupYPositionOffset, transform.position.z);
+
+        transform.position = Vector3.Lerp(transform.position,
+            destinationPosition,
+            Time.deltaTime * GrappleStartupYPositionLerpStrength);
+
+        _momentum = Mathf.Max(0, _momentum - MomentumLossRate * Time.deltaTime * GrappleStartupMomentumLossMod);
+        HandleCollisionMove();
+    }
+}
