+++ Assets/Code/Fsm/GravityFsm/PlayerFsm/PlayerFsm_GroundMove.cs
@@ -1,0 +1,15 @@
+using UnityEngine;
+
+public partial class PlayerFsm
+{
+    private void GroundMoveOnUpdate()
+    {
+        HandleInputMomentumChange();
+        HandleTurning();
+        HandleCollisionMove();
+
+        SetAnimatorMomentum();
+        var speedMod = Mathf.Lerp(0f, GroundMoveMaximumAnimatorSpeedMod, ComputeMomentumWeight());
+        Animator.SetFloat("SpeedMod", speedMod);
+    }
+}
