+++ Assets/Code/Fsm/GravityFsm/PlayerFsm/PlayerFsm_HardLandRoll.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+public partial class PlayerFsm
+{
+    private void HardLandRollOnUpdate()
+    {
+        transform.position += ComputeCollisionMove(transform.forward * (HardLandRollForwardSpeed * Time.deltaTime));
+    }
+}
