+++ Assets/Code/Fsm/GravityFsm/PlayerFsm/PlayerFsm_ImpaleAir.cs
@@ -1,0 +1,18 @@
+using UnityEngine;
+
+public partial class PlayerFsm
+{
+    private void ImpaleAirOnUpdate()
+    {
+        Animator.SetLayerWeight(2, Mathf.Lerp(Animator.GetLayerWeight(2), 1, Time.deltaTime * 90f));
+        Animator.SetLayerWeight(1, 0);
+
+        var speedMod = Mathf.Lerp(0f, GroundMoveMaximumAnimatorSpeedMod, ComputeMomentumWeight());
+        Animator.SetFloat("SpeedMod", speedMod);
+
+        var targetMomentum = _stateEntryMomentum < ImpaleMinimumMomentumAfterOffset
+            ? _momentum
+            : Mathf.Max(_stateEntryMomentum + ImpaleMomentumOffset, ImpaleMinimumMomentumAfterOffset);
+        _momentum = Mathf.Lerp(_momentum, targetMomentum, Time.deltaTime * ImpaleMomentumLerpStrenth);
+    }
+}
