+++ Assets/Code/Fsm/GravityFsm/PlayerFsm/PlayerFsm_LockMomentum.cs
@@ -1,0 +1,11 @@
+using UnityEngine;
+
+public partial class PlayerFsm
+{
+    private void LockMomentumOnUpdate()
+    {
+        Animator.SetFloat("SpeedMod", Mathf.Lerp(0, GroundMoveMaximumAnimatorSpeedMod, ComputeMomentumWeight()));
+        SetAnimatorMomentum();
+        HandleCollisionMove();
+    }
+}
