+++ Assets/Code/Fsm/GravityFsm/PlayerFsm/PlayerFsm_VaultHang.cs
@@ -1,0 +1,9 @@
+public partial class PlayerFsm
+{
+    private void VaultHangOnUpdate()
+    {
+        UpdateLedgePosition(FaceHighLedgeHeight);
+        MoveYOntoLedge(VaultHangLedgeYOffset, VaultHangLedgeLerpStrength);
+        HandleCollisionMove();
+    }
+}
