+++ Assets/Code/Fsm/GravityFsm/PlayerFsm/PlayerFsm_Wallrun.cs
@@ -1,0 +1,14 @@
+using UnityEngine;
+
+public partial class PlayerFsm
+{
+    private void WallrunOnUpdate()
+    {
+        SetAnimatorMomentum();
+        HandleFlankAlignment();
+        HandleCollisionMove();
+
+        transform.position +=
+            ComputeCollisionMove(-_currentFlankWallNormal * (Time.deltaTime * FlankWallVacuumStrength));
+    }
+}
